[#13] refactor : 별점 UI 분리
